refactor(dashboard): share table skeleton rows
