+++ apps/dashboard/src/components/automation/events-skeleton.tsx
@@ -1,31 +1,17 @@
 "use client";
 
 import { Skeleton } from "@notra/ui/components/ui/skeleton";
-import { useId } from "react";
+
+import { TableListSkeletonRows } from "@/components/table-list-skeleton-rows";
 
 export function EventsPageSkeleton() {
-  const id = useId();
   return (
     <div className="space-y-4">
       <Skeleton className="h-10 w-48" />
       <div className="space-y-3">
         <div className="border-border/80 border-b-border/40 bg-muted/80 overflow-hidden rounded-lg border shadow-2xs">
           <div className="bg-background space-y-3 rounded-t-lg p-4">
-            {Array.from({ length: 5 }).map((_, i) => (
-              <div
-                className="flex items-center gap-4 py-2"
-                key={`${id}-row-${i}`}
-              >
-                <Skeleton className="h-8 w-8 rounded-lg" />
-                <Skeleton className="h-4 w-32" />
-                <Skeleton className="h-4 w-24" />
-                <Skeleton className="h-5 w-16 rounded-full" />
-                <Skeleton className="h-4 w-20" />
-                <div className="ml-auto">
-                  <Skeleton className="h-8 w-8 rounded-md" />
-                </div>
-              </div>
-            ))}
+            <TableListSkeletonRows count={5} />
           </div>
         </div>
       </div>
