+++ apps/dashboard/src/types/components/table-list-skeleton.ts
@@ -1,0 +1,3 @@
+export interface TableListSkeletonRowsProps {
+  count: number;
+}
